requirements.txt
PyPDF2>=3.0.0
pillow>=10.0.0
reportlab>=4.0.0
pdf2image>=1.17.0


